Add readme for motivation
its mainly a test to see how many ways we can
deploy next. While keeping a similar set of
primitives across each.

The main aim is to see how similar i can make them,
and compare and contrast.

Ideally it would be a push of a docker file, or a
similar easy primitive