Add readme for motivation
its mianly a test to see how many ways we can deploy next.
While keeping a similar set of primitves acrross each.

The main aim is to see how similar i can make them, and compare and contrast.

Ideally it owudl be a push of a docker file, or a similar easy primitve